Added `execute_binary`, `contains_item` and `contains_item_sorted`. (#3)
Added execute_binary rule. It defines a runnable target that executes another binary target with the specified set of arguments. It wraps a call to the binary in a script with the arguments inside.
Added contains_item and contains_item_sorted to check if an item is contained in a list of arguments, usually an expanded array.
Updated assert_equal fail message argument to be a prefix instead of replacing the error message.
Added a binary target to execute performance tests for contains_item and contains_item_sorted.